+++ scripts/embedding/lvm/lvm_logtalk_pl.ps1
@@ -107,7 +107,7 @@ function Write-Usage-Help() {
 	Write-Output "  -t temporary directory for intermediate files (absolute path; default is an auto-created directory)"
 	Write-Output "  -n name of the generated saved state (default is application)"
 	Write-Output ("  -p library paths file (absolute path; default is " + $p + ")")
-	Write-Output ("  -s settings file (absolute path; default is " + $s + ")")
+	Write-Output ("  -s settings file (absolute path or 'none'; default is " + $s + ")")
 	Write-Output "  -l loader file for the application (absolute path)"
 	Write-Output "  -f copy foreign library files loaded by the application"
 	Write-Output "  -x encrypt the generated logtalk.pl and application.pl files"
@@ -134,7 +134,7 @@ function Check-Parameters() {
 		Exit
 	}
 
-	if (($s -ne "") -and (-not(Test-Path $s))) {
+	if (($s -ne "") -and ($s -ne "none") -and (-not(Test-Path $s))) {
 		Write-Output ("The " + $s + " settings file does not exist!")
 		Start-Sleep -Seconds 2
 		Exit
@@ -229,7 +229,7 @@ if ($c -eq $true) {
 	Copy-Item -Path $p -Destination ($t + '\paths_lgt.pl')
 }
 
-if ($s -eq "") {
+if (($s -eq "") -or ($s -eq "none")) {
 	Get-Content -Path lvm.pl,
 		paths_*.pl,
 		expanding*_lgt.pl,